requirements.txt
langgraph
langchain
pillow==11.2.1
streamlit
dotenv
langchain-community
python-dotenv
huggingface-hub
langchain-openai
boto3